pyproject.toml
[tool.poetry]
name = "mmwave-gesture-recognition"
version = "0.1.0"
description = "mmWave Gesture Recognition"
authors = ["Loki <yongxiang.zhan@outlook.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9 <3.10"
keras = "^2.15.0"
matplotlib = "^3.8.2"
numpy = "^1.26.4"
pyserial = "^3.5"
tensorflow = "^2.15.0"
PyQt5 = "^5.15.11"
rich = "^13.7.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

[tool.poetry.scripts]
mmwave-gesture-recognition = "src.main:main"
